Fix Getting Self Agent Handle in perception_module.py
Because in newest code we get the handle representing bot itself from EntityReocorder, not SpaceMap, the old code will fail. Now we add a helper function to get entity recorder and recover this functionality.
Note that we get the EntityRecorder by passing the name of space map. In fact the "map name" should represent the scene that the bot lives in. And in one scene it corresponding to one space map and one entity recorder. But since we don't have anything such as "SceneNode" currently we just use the SpaceMapNode to represent both space map and entity recorder, using the map name as the content of handle.